+++ src/gwt/src/org/rstudio/studio/client/workbench/prefs/PrefsConstants.java
@@ -1168,13 +1168,22 @@ public interface PrefsConstants extends com.google.gwt.i18n.client.Messages {
     String consoleLimitVariableLabel();
 
     /**
-     * Translated "Limit output line length to:".
+     * Translated "Truncate lines to maximum length (characters)".
      *
-     * @return translated "Limit output line length to:"
+     * @return translated "Truncate lines to maximum length (characters)"
      */
-    @DefaultMessage("Limit output line length to:")
+    @DefaultMessage("Truncate lines to maximum length (characters)")
     @Key("consoleLimitOutputLengthLabel")
     String consoleLimitOutputLengthLabel();
+
+    /**
+     * Translated "Number of lines to show in console history:".
+     *
+     * @return translated "Number of lines to show in console history:"
+     */
+    @DefaultMessage("Number of lines to show in console history:")
+    @Key("consoleMaxLinesLabel")
+    String consoleMaxLinesLabel();
 
     /**
      * Translated "ANSI Escape Codes:".