+++ NEWS.md
@@ -8,6 +8,7 @@
 - Improved legibility of highlighted code when RStudio debugger is active
 - Added a user preference to disable showing the splash screen at startup (#15945)
 - The splash screen now closes when clicked with the mouse (#15614)
+- The default console buffer size has been increased to 10000 lines (#16111)
 
 #### Posit Workbench